% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRB.R
\name{HRB}
\alias{HRB}
\title{Hierarchically Regularized Entropy Balancing}
\usage{
HRB(x, treat, group1, group2 = NA, outcome, iterations = 1000)
}
\arguments{
\item{x}{covariates.}

\item{treat}{treatment indicator vector.}

\item{group1}{see Details.}

\item{group2}{see Details, Default: NA.}

\item{outcome}{outcome vector.}

\item{iterations}{iteration time in optimization problem, Default: 1000.}
}
\value{
a HRB object with the following attributes:
\itemize{
\item{AT:}{ the estimate of average treatment effect in group1 (i.e, \eqn{E(Y(group1))}).}
\item{weight:}{ the estimated Mahalanobis balancing weight.}
\item{GMIM:}{ Generalized Multivariate Imbalance Measure that defines in our paper.}
}
}
\description{
Hierarchically Regularized Entropy Balancing expands the feature space
by including higher-order terms (including interactions) of covariates
and then achieves approximate balance on the expanded features.
effect.
}
\details{
\code{group1} and \code{group0}
\itemize{
\item To estimate \eqn{E(Y (1))} (average treatment effect for group 1),
you need to set \code{group1} = 1 and ignore \code{group2}. Similarly,
To estimate \eqn{E(Y (0))} (average treatment effect for group 0),
you need to set \code{group1} = 0 and ignore \code{group2}.

\item To estimate average treatment effect on the control group \eqn{E(Y (1) | T = 0)},
you need to set \code{group1} = 1 and \code{group2} = 0. Similarly, To estimate
average treatment effect on the treated group \eqn{E(Y (0) | T = 1)},
you need to set \code{group1} = 0 and \code{group2} = 1.

\item This function is feasible when there are more than two groups.
}
}
\examples{
## estimating ATE##
set.seed(0521)
data <- si.data()


}
