% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.data.R
\name{si.data}
\alias{si.data}
\title{Generate data}
\usage{
si.data(sample.size = 500, dimension = 10)
}
\arguments{
\item{sample.size}{sample size for data, Default: 500,}

\item{dimension}{dimension for covariate, it should be larger than 10, Default: 10.}
}
\value{
data object with the following attributes:
\itemize{
\item{X:}{ covariates.}
\item{Tr:}{ treatment indicator.}
\item{Y:}{ observed outcome.}
}
}
\description{
Generate data for testing our main function \code{MB} and \code{hdMB}..
}
\details{
We first generate covariates X from standard normal distribution function.
Then we will use the following propensity score model with the binomial
distribution to generate treatment indicator T:

\eqn{Pr(T = 1 | X) = 1/ (1 + exp(X_1 + ... + X_{10}))}

Finally, the outcome model will be generated by
\eqn{Y = (X_1 + ... + X_{10}) + noise},
where noise follows standard normal distribution.
}
\examples{
data <- si.data()
}
