% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{soft}
\usage{
soft(vt, lambda, norm = "square-l2")
}
\arguments{
\item{vt}{input for proximal operator}

\item{lambda}{threshold}

\item{norm}{norm type}
}
\value{
soft_value output for proximal operator
}
\description{
proximal operator for l_1 norm and l_2 norm
}
