% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholMB.R
\name{CholMB}
\alias{CholMB}
\title{CholMB}
\usage{
CholMB(covariate, treat, group1, group2 = NA, method = "MB")
}
\arguments{
\item{covariate}{covariate.}

\item{treat}{treatment indicator vector.}

\item{group1}{see Details.}

\item{group2}{see Details. Default: NA.}

\item{method}{a string that takes values in {"MB", "MB2"}, Default: 'MB'.}
}
\description{
Function to decompose weighting matrix.
}
\details{
\code{group1} and \code{group0}
\itemize{
\item To estimate \eqn{E(Y (1))} (average treatment effect for group 1),
you need to set \code{group1} = 1 and ignore \code{group2}. Similarly,
To estimate \eqn{E(Y (0))} (average treatment effect for group 0),
you need to set \code{group1} = 0 and ignore \code{group2}.

\item To estimate average treatment effect on the control group \eqn{E(Y (1) | T = 0)},
you need to set \code{group1} = 1 and \code{group2} = 0. Similarly, To estimate
average treatment effect on the treated group \eqn{E(Y (0) | T = 1)},
you need to set \code{group1} = 0 and \code{group2} = 1.

\item This function is feasible when there are more than two groups.
}
}
