% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRB.R
\name{HRB}
\alias{HRB}
\title{Hierarchically Regularized Entropy Balancing}
\usage{
HRB(
  covariate,
  treat,
  group1,
  group2 = NULL,
  outcome,
  alpha.space = c(0.1, 0.05, 0.01, 0.005, 0.001),
  iterations = 1000,
  convergence = 10^{
     -10
 },
  cv.fold = 5,
  second.moment = TRUE,
  third.moment = TRUE,
  interact = TRUE
)
}
\arguments{
\item{covariate}{covariates.}

\item{treat}{treatment indicator.}

\item{group1}{see Details.}

\item{group2}{see Details, Default: NULL.}

\item{outcome}{outcome.}

\item{alpha.space}{alpha.space is}

\item{iterations}{iteration time in optimization problem, Default: 1000.}

\item{convergence}{Criteria for determining convergence of optimization problems. Default: 10^{-10}.}

\item{cv.fold}{cv.fold is}

\item{second.moment}{second.moment is}

\item{third.moment}{third.moment is}

\item{interact}{interact is}
}
\value{
a HRB object with the following attributes:
\itemize{
\item{AT:}{ the estimate of average treatment effect in group1 (i.e, \eqn{E(Y(group1))}).}
\item{weight:}{ the estimated univariate balancing weight.}
\item{GMIM:}{ Generalized Multivariate Imbalance Measure that defines in our paper.}
\item{MAE:}{ Mean Absolute Error.}
\item{alpha:}{ the tuning parameter we choose.}
}
}
\description{
Hierarchically Regularized Entropy Balancing is a
approximate balancing method that expands the feature space by including
higher-order terms of covariates while imposes
ridge penalties with a hierarchical structure on the higher-order terms.
}
\examples{
# estimating ATE
set.seed(0521)
data <- si.data()
result1 <- HRB(x = data$X, treat = data$Tr, group1 = 1, outcome = data$Y)
result2 <- HRB(x = data$X, treat = data$Tr, group1 = 0, outcome = data$Y)

# an estimate of ATE
result1$AT - result2$AT

# estimating ATC
result3 <- MB(x = data$X, treat = data$Tr, group1 = 1, group2 = 0, outcome = data$Y, method = "MB")

# an estimate of ATC
result3$AT - mean(data$Y[data$Tr == 0])
}
