% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB.R
\name{MB}
\alias{MB}
\title{Mahalanobis balancing}
\usage{
MB(
  covariate,
  treat,
  group1,
  group2 = NULL,
  outcome,
  opti.method = c("proximalC", "proximal", "BFGS"),
  method = c("MB", "MB2", "kernelMB"),
  delta.space = c(0.1, 0.05, 0.01, 0.005, 0.001),
  iterations = 1000,
  convergence = 1e-08,
  rate = 10
)
}
\arguments{
\item{covariate}{Covariates, it should be a n by d matrix where n is sample size and d is dimension.}

\item{treat}{The vector of treatment assignments, it should be a vector of length n.}

\item{group1}{group1 is a number, see Details.}

\item{group2}{group2 is a number, see Details, Default: NULL.}

\item{outcome}{The outcome vector, it should be a vector of length n.}

\item{opti.method}{The optimization method that will be applied to optimize loss function. It should takes values in {"BFGS", "proximal", "proximalC"}. Default: "proximalC".}

\item{method}{Method to calculate weighting matrix. It should takes values in {"MB", "MB2", "kernelMB"}. See Details. Default: "MB".}

\item{delta.space}{Tuning parameter in balancing. See Details. Default: c(1e-1, 0.05, 1e-2, 0.005, 1e-3).}

\item{iterations}{The maximum number of iteration times in optimization problem, Default: 1000.}

\item{convergence}{The absolute tolerance used to determine a stopping criteria for the optimization problems. Default: 10^{-8}.}

\item{rate}{The parameter in Tikhonov-regularized Newton update. Default: 10.}
}
\value{
a MB object with the following attributes:
\itemize{
\item{AT:}{ the estimate of average treatment effect in group1 (i.e, \eqn{E(Y(group1))}).}
\item{weight:}{ the estimated Mahalanobis balancing weight.}
\item{GMIM:}{ Generalized Multivariate Imbalance Measure that defines in our paper.}
\item{delta:}{ the tuning parameter we choose.}
}
}
\description{
Mahalanobis balancing (Dai, Y., & Yan, Y. (2022).) is a multivariate perspective of
approximate covariate balancing method to estimate average treatment
effect.
}
\details{
\code{group1} and \code{group0}
\itemize{
\item To estimate \eqn{E(Y (1))} (average treatment effect for group 1),
you need to set \code{group1} = 1 and ignore \code{group2}. Similarly,
To estimate \eqn{E(Y (0))} (average treatment effect for group 0),
you need to set \code{group1} = 0 and ignore \code{group2}.

\item To estimate average treatment effect on the control group \eqn{E(Y (1) | T = 0)},
you need to set \code{group1} = 1 and \code{group2} = 0. Similarly, To estimate
average treatment effect on the treated group \eqn{E(Y (0) | T = 1)},
you need to set \code{group1} = 0 and \code{group2} = 1.

\item This function is feasible when there are more than two groups.
}

\code{method} can be a valid string, including
\itemize{
\item "MB": We choose the weighting matrix \eqn{{W}_1=[diag(\hat{\Sigma})]^{-1}}
where \eqn{\hat{\Sigma}} denotes sample covariance matrix.
\item "MB2": We choose the weighting matrix \eqn{{W}_2={[\hat{\Sigma}]^{-1}}}
where \eqn{\hat{\Sigma}} denotes sample covariance matrix.
\item "kernelMB": Firstly, we modify our covariate to \eqn{X_i* = (\Phi(X_1,X_i), ..., \Phi(X_n,X_i))},
then we apply method "MB" to produce Mahalanobis balancing weights.
}

\code{delta.space} grid of values for the tuning parameter, a vector of
candidate values for the degree of multivariate approximate covariate balance.
The default is c(1e-1, 0.05, 1e-2, 0.005, 1e-3).
}
\examples{
# estimating ATE
set.seed(0521)
data <- si.data()
X <- data$X
treat <- data$Tr
Y <- data$Y

result1 <- MB(covariate = X, treat = treat, group1 = 1, outcome = Y)
result2 <- MB(covariate = X, treat = treat, group1 = 0, outcome = Y)

# an estimate of ATE
result1$AT - result2$AT

# estimating ATC
result3 <- MB(covariate = X, treat = treat, group1 = 1, group2 = 0, outcome = Y)

# an estimate of ATC
result3$AT - mean(data$Y[data$Tr == 0])

}
\references{
Dai, Y., & Yan, Y. (2022). Mahalanobis balancing: a multivariate perspective on approximate covariate balancing. arXiv preprint arXiv:2204.13439.
}
